Playbook to create local user on Raspberry
This playbook creates a local user, copies the playbooks/files/authorized_key files to the user's directory that is used for authentication (the user does not have a password!).
The user will also be added to sudoers to make sure that this account can be used to install all software needed.