Added playbook to fetch settings from the Airplay devices
at the moment this playbook downloads only the authorized_keys file
from one of the Airplay speakers to the directory on the local machine
in which this file will be expected by other playbooks.
We download this file only from one of the airplay speakers because
we assume that these devices have the same authorized_keys file.